+++ lib/src/features/dashboard/presentation/screen/dashboard_screen.dart
@@ -1,4 +1,5 @@
 import 'package:fluent_ui/fluent_ui.dart';
+import 'package:gui/src/core/common/widgets/app_layout.dart';
 import 'package:gui/src/features/blockchain_get_info/presentation/pages/blockchain_info_section.dart';
 
 class DashboardScreen extends StatelessWidget {
@@ -6,23 +7,23 @@ class DashboardScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return NavigationView(
-      appBar: NavigationAppBar(
-        automaticallyImplyLeading: false,
-      ),
-      content: PopScope(
-        canPop: false,
-        child: Center(
-          child: Column(
-            children: [
-              const Text(
-                'DashboardPage',
-                style: TextStyle(
-                  color: Colors.black,
+    return AppLayout(
+      isDashboard: true,
+      content: NavigationView(
+        content: PopScope(
+          canPop: false,
+          child: Center(
+            child: Column(
+              children: [
+                const Text(
+                  'DashboardPage',
+                  style: TextStyle(
+                    color: Colors.black,
+                  ),
                 ),
-              ),
-              BlockchainInfoSection(),
-            ],
+                BlockchainInfoSection(),
+              ],
+            ),
           ),
         ),
       ),